requirements.txt
pygelbooru
Mastodon.py
urllib
pillow
